+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Minimalistic fork of [mock-server/mockserver](https://github.com/mock-server/moc
 * Copied over original code of the following modules with minimalistic dependencies:
   * client-java -> client
   * core -> core
-  * netty -> standalone
+  * netty -> server
 * Brought code into a maintainable state
   * Removed about 75% of the original code
   * Removed everything that is not core functionality: